Rewrite pytdml type according to standard TrainingDML-AI jsonschema, and add some unitest
