+++ tests/data/WHU-building.json
@@ -1,0 +1,67 @@
+{
+  "type": "AI_EOTrainingDataset",
+  "id": "whu-building",
+  "name": "WHU-Building Dataset-Building Change Detection Dataset",
+  "description": "This dataset covers an area where a 6.3-magnitude earthquake has occurred in February 2011 and rebuilt in the following years. This dataset consists of aerial images obtained in April 2012 that contains 12796 buildings in 20.5 km2 (16077 buildings in the same area in 2016 dataset).",
+  "license": "CC BY-SA 4.0",
+  "version": "1.0",
+  "amountOfTrainingData": 1,
+  "createdTime": "2011-01-01",
+  "providers": [
+    "Wuhan University"
+  ],
+  "extent": [
+    1560160,
+    5176338.4,
+    1566661.4,
+    5179409.2
+  ],
+  "classes": [
+    {
+      "building": 1
+    }
+  ],
+  "numberOfClasses": 1,
+  "bands": [
+    "red",
+    "green",
+    "blue"
+  ],
+  "imageSize": "32507x15354",
+  "tasks": [
+    {
+      "type": "AI_EOTask",
+      "id": "whu-building-task",
+      "description": "Building change detection based on aerial images",
+      "taskType": "Change Detection"
+    }
+  ],
+  "data": [
+    {
+      "type": "AI_EOTrainingData",
+      "id": "0",
+      "extent": [
+        1560160,
+        5176338.4,
+        1566661.4,
+        5179409.2
+      ],
+      "dataTime": [
+        "2012-01-01",
+        "2016-01-01"
+      ],
+      "dataURL": [
+        "before/before.tif",
+        "after/after.tif"
+      ],
+      "numberOfLabels": 1,
+      "labels": [
+        {
+          "type": "AI_PixelLabel",
+          "imageURL": ["change label/change_label.tif"],
+          "imageFormat": ["TIFF"]
+        }
+      ]
+    }
+  ]
+}
